readyup|refactor: Move the JSON Schema builder beside the schemas it reads
The module deriving readyup's published JSON Schema documents sits with the payload schemas it reads, so its suite sits beside it too.
